+++ frontend/e2e/common/record-api-mock.ts
@@ -2,6 +2,8 @@ import { Page } from "@playwright/test";
 import { DatasetData, mockFeedbackTaskDataset } from "./dataset-api-mock";
 import {
   mockQuestion,
+  mockQuestionWith12Ranking,
+  mockQuestionWithRating,
   mockQuestionLongAndShortQuestions,
 } from "./question-api-mock";
 import { mockFields } from "./field-api-mock";
@@ -91,6 +93,45 @@ export const recordTwo = {
   updated_at: "2023-07-18T07:43:38",
 };
 
+const recordFor12rankingA = {
+  id: "1da11112-69ac-4cc9-947e-c8293243510a",
+  fields: {
+    text: "First Record",
+  },
+  metadata: {},
+  external_id: null,
+  responses: [],
+  suggestions: [],
+  inserted_at: "2023-07-26T12:15:02",
+  updated_at: "2023-07-26T12:15:02",
+};
+
+const recordFor12rankingB = {
+  id: "1da11112-69ac-4cc9-947e-c8293243510b",
+  fields: {
+    text: "Second record",
+  },
+  metadata: {},
+  external_id: null,
+  responses: [],
+  suggestions: [],
+  inserted_at: "2023-07-26T12:15:02",
+  updated_at: "2023-07-26T12:15:02",
+};
+
+const recordForRating = {
+  id: "0203fc47-e30a-4f97-8f13-12bb816a3059",
+  fields: {
+    text: "Rate me",
+  },
+  metadata: {},
+  external_id: null,
+  responses: [],
+  suggestions: [],
+  inserted_at: "2023-07-21T09:23:20",
+  updated_at: "2023-07-21T09:23:20",
+};
+
 export const mockRecord = async (
   page: Page,
   { datasetId, workspaceId }: DatasetData
@@ -162,33 +203,19 @@ export const mockRecordForLongAndShortQuestion = async (
   return recordOne;
 };
 
-export const mockDiscardRecord = async (page: Page, recordId: string) => {
+export const mockRecordResponses = async (
+  page: Page,
+  recordId: string,
+  status: "submitted" | "discarded"
+) => {
   await page.route(
     `*/**/api/v1/records/${recordId}/responses`,
-    async (route) => {
+    async (route, request) => {
       await route.fulfill({
         json: {
           id: recordId,
-          values: {},
-          status: "discarded",
-          user_id: "3e760b76-e19a-480a-b436-a85812b98843",
-          inserted_at: "2023-07-28T14:45:37",
-          updated_at: "2023-07-28T14:45:37",
-        },
-      });
-    }
-  );
-};
-
-export const mockSubmitRecord = async (page: Page, recordId: string) => {
-  await page.route(
-    `*/**/api/v1/records/${recordId}/responses`,
-    async (route) => {
-      await route.fulfill({
-        json: {
-          id: recordId,
-          values: {},
-          status: "submitted",
+          values: request.postDataJSON().values,
+          status,
           user_id: "3e760b76-e19a-480a-b436-a85812b98843",
           inserted_at: "2023-07-28T14:45:37",
           updated_at: "2023-07-28T14:45:37",
@@ -231,6 +258,60 @@ export const mockDraftRecord = async (page: Page, recordId: string) => {
       });
 
       debounce.stop();
+    }
+  );
+};
+
+export const mockRecordWith12Ranking = async (
+  page: Page,
+  { datasetId, workspaceId }: DatasetData
+) => {
+  await mockFeedbackTaskDataset(page, { datasetId, workspaceId });
+
+  await mockQuestionWith12Ranking(page, datasetId);
+
+  await mockFields(page, datasetId);
+
+  await page.route(
+    `*/**/api/v1/me/datasets/${datasetId}/records?include=responses&include=suggestions&offset=0&limit=10&response_status=missing&response_status=draft`,
+    async (route) => {
+      await route.fulfill({
+        json: {
+          items: [recordFor12rankingA, recordFor12rankingB],
+        },
+      });
+    }
+  );
+  await page.route(
+    `*/**/api/v1/me/datasets/${datasetId}/records?include=responses&include=suggestions&offset=2&limit=10&response_status=missing&response_status=draft`,
+    async (route) => {
+      await route.fulfill({
+        json: {
+          items: [],
+        },
+      });
+    }
+  );
+};
+
+export const mockRecordWithRating = async (
+  page: Page,
+  { datasetId, workspaceId }: DatasetData
+) => {
+  await mockFeedbackTaskDataset(page, { datasetId, workspaceId });
+
+  await mockQuestionWithRating(page, datasetId);
+
+  await mockFields(page, datasetId);
+
+  await page.route(
+    `*/**/api/v1/me/datasets/${datasetId}/records?include=responses&include=suggestions&offset=0&limit=10&response_status=missing&response_status=draft`,
+    async (route) => {
+      await route.fulfill({
+        json: {
+          items: [recordForRating],
+        },
+      });
     }
   );
 };