+++ tests/integration/client/feedback/training/test_trainer.py
@@ -35,6 +35,7 @@ from argilla.client.feedback.schemas import (
 from argilla.client.feedback.training import ArgillaTrainer
 from argilla.client.feedback.training.schemas import (
     TrainingTask,
+    TrainingTaskForQuestionAnsweringFormat,
     TrainingTaskForTextClassification,
     TrainingTaskForTextClassificationFormat,
     TrainingTaskMapping,
@@ -143,13 +144,6 @@ def test_prepare_for_training_text_classification_with_formatting_func(
         else:
             return None
 
-    def correct_formatting_func(sample):
-        data = wrong_formatting_func(sample)
-        if data:
-            return (data["text"], data["label"])
-        else:
-            return None
-
     with pytest.raises(
         ValueError,
         match=re.escape(
@@ -161,8 +155,108 @@ def test_prepare_for_training_text_classification_with_formatting_func(
         trainer.update_config(num_iterations=1)
         trainer.train(__OUTPUT_DIR__)
 
+    def correct_formatting_func(sample):
+        data = wrong_formatting_func(sample)
+        if data:
+            yield (data["text"], data["label"])
+        else:
+            yield None
+
     task = TrainingTask.for_text_classification(correct_formatting_func)
     trainer = ArgillaTrainer(dataset=dataset, task=task, framework=framework)
+    trainer.update_config(num_iterations=1)
+    trainer.train(__OUTPUT_DIR__)
+
+    def correct_formatting_func_with_yield(sample):
+        data = wrong_formatting_func(sample)
+        if data:
+            yield (data["text"], data["label"])
+        else:
+            yield None
+
+    task = TrainingTask.for_text_classification(correct_formatting_func_with_yield)
+    trainer = ArgillaTrainer(dataset=dataset, task=task, framework=framework)
+    trainer.update_config(num_iterations=1)
+    trainer.train(__OUTPUT_DIR__)
+
+
+@pytest.mark.usefixtures(
+    "feedback_dataset_guidelines",
+    "feedback_dataset_fields",
+    "feedback_dataset_questions",
+    "feedback_dataset_records",
+)
+def test_question_answering_with_formatting_func(
+    feedback_dataset_fields, feedback_dataset_questions, feedback_dataset_records, feedback_dataset_guidelines
+):
+    dataset = FeedbackDataset(
+        guidelines=feedback_dataset_guidelines,
+        fields=feedback_dataset_fields,
+        questions=feedback_dataset_questions,
+    )
+    dataset.add_records(records=feedback_dataset_records * 5)
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            f"formatting_func must return {TrainingTaskForQuestionAnsweringFormat.__annotations__['format']}, not <class 'dict'>"
+        ),
+    ):
+        task = TrainingTask.for_question_answering(lambda x: {})
+        trainer = ArgillaTrainer(dataset=dataset, task=task, framework="transformers")
+        trainer.update_config(num_iterations=1)
+        trainer.train(__OUTPUT_DIR__)
+
+    def formatting_func(sample):
+        responses = []
+        question = sample["label"]
+        context = sample["text"]
+        for answer in sample["question-1"]:
+            if not all([question, context, answer["value"]]):
+                continue
+            responses.append((question, context, answer["value"]))
+        return responses
+
+    task = TrainingTask.for_question_answering(formatting_func)
+    trainer = ArgillaTrainer(dataset=dataset, task=task, framework="transformers")
+    trainer.update_config(num_iterations=1)
+    trainer.train(__OUTPUT_DIR__)
+
+    def formatting_func_with_yield(sample):
+        question = sample["label"]
+        context = sample["text"]
+        for answer in sample["question-1"]:
+            if not all([question, context, answer["value"]]):
+                continue
+            yield question, context, answer["value"]
+
+    task = TrainingTask.for_question_answering(formatting_func_with_yield)
+    trainer = ArgillaTrainer(dataset=dataset, task=task, framework="transformers")
+    trainer.update_config(num_iterations=1)
+    trainer.train(__OUTPUT_DIR__)
+
+
+@pytest.mark.usefixtures(
+    "feedback_dataset_guidelines",
+    "feedback_dataset_fields",
+    "feedback_dataset_questions",
+    "feedback_dataset_records",
+)
+def test_question_answering_without_formatting_func(
+    feedback_dataset_fields, feedback_dataset_questions, feedback_dataset_records, feedback_dataset_guidelines
+):
+    dataset = FeedbackDataset(
+        guidelines=feedback_dataset_guidelines,
+        fields=feedback_dataset_fields,
+        questions=feedback_dataset_questions,
+    )
+    dataset.add_records(records=feedback_dataset_records * 5)
+
+    task = TrainingTask.for_question_answering(
+        question=dataset.field_by_name("label"),
+        context=dataset.field_by_name("text"),
+        answer=dataset.question_by_name("question-1"),
+    )
+    trainer = ArgillaTrainer(dataset=dataset, task=task, framework="transformers")
     trainer.update_config(num_iterations=1)
     trainer.train(__OUTPUT_DIR__)